calculate hash, for message store
